requirements.txt
tensorflow-gpu==2.5.1
opencv==4.5.2
tqdm